+++ frontend/src/app/_homeComponents/Hero.tsx
